Add book constructor, array modifier, HTML outline
The basic way I see this working is by outputting the contents of the
Library Array into the table. The script currently does not create an
array of objects but instead an array of the information contained
within the objects as they are created, so that may need to change.