Adjust script to handle objects instead of info
Again, I am not sure how I want this handled but I think for a cleaner
script it makes more sense to keep the Array full of objects instead of
just the return of the info function. I suspect this will make it easier
to do the DOM manipulation to add the book info into the table (or
whatever I decide to put it into).