Add Script for adding Books, adjust table layout
The table is now laid out more sensibly with a thead and tbody. The
script now adds a new row with the information gathered from the
information gathered from the prompts. This of course will change to
some kind of form, but for now I wanted to make sure the script was
working as intended. I would like to see if there is a clean way to add
a scope attribute to the new rows as they are added. I also need to
personally add the scope to the headings.