Change display function to match project guideline
My display function was not looping through the array and displaying
items on the page but instead directly taking new books when they were
being added instead. In order to better align with the guidelines in the
project, the display function now loops over each object in the array
when called, adds a row and the appropriate number of cells, then
populates those cells with the relevant information.