Add delete button, add data-* to tie book to index
The basic idea here was to change the way the function for adding books
worked so that when a new book was created and added to the library it's
index would be captured and then added as a data attribute to the table
row element. Then I added a section of code that makes it so that a
delete button is appended to a new cell in that table. Now I just have
to add the code that makes it so that when the button is pressed, it
deletes the row that it is a part of and then removes that book from the
library array by associating it's index with the data attribute.