Add read status update functionality script
When a book is added, a button is generated that allows the user to
switch between having the book be read or not read. This works by using
a prototypal function that simply updates the value of the read status.
It acquires this by using the same data attribute for the row, then
using that index to find the book that the button belongs to. I can see
now how having these data attributes and prototypal functions can be
useful since any book that is created by the user will have all of this
functionality and I don't have to write insanely long functions to
handle all of this behavior.